refactor: [FE] 컴포넌트 섹션별로 분리 공통 - common 나머지는 section별로 폴더를 만들어서 분리
