+++ frontEnd/src/components/common/Header.tsx
@@ -1,0 +1,8 @@
+export default function Header() {
+  return (
+    <header className="flex items-center justify-start">
+      <img src="/main.png" alt="logo" width="50px" />
+      <div className="font-bold font-Pretendard text-[1.8vw]">AlgoITNi</div>
+    </header>
+  );
+}
